Add Tailwind CSS configuration, update Next.js config for image handling, and introduce new SVG icons
- Added tailwind.config.js for Tailwind CSS setup.
- Updated next.config.ts to allow images from Unsplash.
- Introduced new SVG icons for Apple and Google branding.
- Added various image assets to the public/images directory for use in the application.
- Enhanced layout.tsx to include a new font (Assistant) for improved typography.
- Refactored components to utilize new icons and images, ensuring consistency with design specifications.
