+++ src/components/HomeWithSuspense.tsx
@@ -1,0 +1,31 @@
+"use client";
+
+import { Suspense } from "react";
+import { HomeContent } from "./HomeContent";
+
+interface LoadingFallbackProps {
+  message?: string;
+}
+
+function LoadingFallback({ message = "Loading..." }: LoadingFallbackProps) {
+  return (
+    <div className="min-h-screen bg-gray-50 py-8">
+      <div className="max-w-7xl mx-auto px-4">
+        <div className="flex items-center justify-center h-64">
+          <div className="text-center">
+            <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-blue-500 mx-auto mb-4"></div>
+            <p className="text-gray-600">{message}</p>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export function HomeWithSuspense() {
+  return (
+    <Suspense fallback={<LoadingFallback />}>
+      <HomeContent />
+    </Suspense>
+  );
+}
